requirements.txt
pyaml==20.4.0
scikit-learn==0.24.2
pandas==1.2.2
numpy==1.19.5 
catboost==0.25.1
shap==0.36
matplotlib==3.3.4
streamlit==0.86.0
yellowbrick==1.3.post1